moving potion class to separate package
